gpt/src/backend-python/requirements.txt
Flask==2.1.2
Flask-Cors==3.0.10
psycopg2-binary==2.9.3